Rewrote "every"
