+++ src/modern-async.mjs
@@ -8,8 +8,6 @@ export { default as delay } from './delay.mjs'
 export { default as delayCancellable } from './delayCancellable.mjs'
 export { default as Delayer } from './Delayer.mjs'
 export { default as every } from './every.mjs'
-export { default as everyLimit } from './everyLimit.mjs'
-export { default as everySeries } from './everySeries.mjs'
 export { default as filter } from './filter.mjs'
 export { default as filterGenerator } from './filterGenerator.mjs'
 export { default as filterLimit } from './filterLimit.mjs'
